requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.4
future==0.17.1
idna==2.8
MouseInfo==0.0.4
Pillow==6.1.0
PyAutoGUI==0.9.47
pycparser==2.19
pygame==1.9.6
PyGetWindow==0.0.6
PyMsgBox==1.0.7
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.22
PySocks==1.7.0
python-telegram-bot==11.1.0
PyTweening==1.0.3
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
urllib3==1.25.3
